requirements.txt
fastapi
uvicorn
torch
numpy
pillow
